Add more complex formatting/style pre-commit hooks
This adds `black`, `isort`, and `prettier` for automatic formatting and
`codespell` for spellchecking. I set the line width to 100 (wider than
`baseline_bots`) because the traditional 80-character limit is much
smaller than modern screens. I made the one spelling correction that was
needed and set another existing word to be ignored.