Modify `get_prev_m_phase()` to always return a value
The method relies on game history for its data, but there is no history
when a game starts. To still be able to calculate stance at the
beginning of a game, similar data is extracted from the game's current
state if no usable game history is available.

I also made some minor code improvements while I was improving the
method.