Replace `numpy.random()` with stdlib's `random` module
This was the final usage of `numpy` in the library, so I replaced it
with the very similar method from the Python standard library. As part
of this change, I added an internal instance of `Random` to
`ActionBasedStance` to segregate this random number generation from
Python's global generator. This will make reproducibility and testing
easier.

Now that `numpy` is no longer used directly, I removed it as a direct
dependency.